+++ src/it/scala/org/renci/cam/test/PredicateServiceTest.scala
@@ -1,4 +1,4 @@
-package org.renci.cam
+package org.renci.cam.test
 
 import java.io.StringReader
 import java.nio.file.{Files, Paths}
