+++ src/it/scala/org/renci/cam/test/QueryServiceTest.scala
@@ -1,4 +1,4 @@
-package org.renci.cam
+package org.renci.cam.test
 
 import java.nio.file.{Files, Paths}
