+++ src/test/scala/org/renci/cam/test/QueryServiceTest.scala
@@ -1,0 +1,5 @@
+package org.renci.cam.test
+
+object QueryServiceTest {
+
+}
